refactor: split image resolver per-source, fix debug overlay contrast
Splits image-sources.ts/image-resolver.ts into a shared SourceResolver
protocol plus one file per NASA feed (source-resolver-dscovrearth.ts,
source-resolver-sdosun.ts), each owning its own TTL/base URL. Renames
DateNav to DateNavigation. Darkens debug-overlay background so green
text stays readable on light themes.
